requirements.txt
flask
flask-cors
openai
pinecone-client
gunicorn
